+++ client/test_helpers.go
@@ -45,8 +45,13 @@ type TestAccountRetriever struct {
 }
 
 // GetAccount implements AccountRetriever.GetAccount
-func (t TestAccountRetriever) GetAccount(_ Context, addr sdk.AccAddress) (Account, error) {
-	acc, ok := t.Accounts[addr.String()]
+func (t TestAccountRetriever) GetAccount(clientCtx Context, addr sdk.AccAddress) (Account, error) {
+	addrStr, err := clientCtx.AddressCodec.BytesToString(addr)
+	if err != nil {
+		return nil, err
+	}
+
+	acc, ok := t.Accounts[addrStr]
 	if !ok {
 		return nil, fmt.Errorf("account: account %s not found", addr)
 	}
@@ -65,8 +70,13 @@ func (t TestAccountRetriever) GetAccountWithHeight(clientCtx Context, addr sdk.A
 }
 
 // EnsureExists implements AccountRetriever.EnsureExists
-func (t TestAccountRetriever) EnsureExists(_ Context, addr sdk.AccAddress) error {
-	_, ok := t.Accounts[addr.String()]
+func (t TestAccountRetriever) EnsureExists(clientCtx Context, addr sdk.AccAddress) error {
+	addrStr, err := clientCtx.AddressCodec.BytesToString(addr)
+	if err != nil {
+		return err
+	}
+
+	_, ok := t.Accounts[addrStr]
 	if !ok {
 		return fmt.Errorf("ensureExists: account %s not found", addr)
 	}
@@ -74,8 +84,13 @@ func (t TestAccountRetriever) EnsureExists(_ Context, addr sdk.AccAddress) error
 }
 
 // GetAccountNumberSequence implements AccountRetriever.GetAccountNumberSequence
-func (t TestAccountRetriever) GetAccountNumberSequence(_ Context, addr sdk.AccAddress) (accNum, accSeq uint64, err error) {
-	acc, ok := t.Accounts[addr.String()]
+func (t TestAccountRetriever) GetAccountNumberSequence(clientCtx Context, addr sdk.AccAddress) (accNum, accSeq uint64, err error) {
+	addrStr, err := clientCtx.AddressCodec.BytesToString(addr)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	acc, ok := t.Accounts[addrStr]
 	if !ok {
 		return 0, 0, fmt.Errorf("accountNumberSequence: account %s not found", addr)
 	}